Clear the warnings the release build reports
Six warnings across the release matrix, all ours:

  - CMP0135 was unset, so FetchContent kept the archives' own
    timestamps and a changed URL would not rebuild what depends on the
    contents.  Set it where the policy exists.
  - index_builder dropped the fread return value when reading a posting
    file back out of the .tmp; a short read now reports and fails like
    the other I/O errors there.  The buffer it lands in is the posting
    file, so it is named that.
  - test_parallel_sort narrowed the mt19937 result to uint32_t inside a
    braced initialiser.
  - test_spaced_seed compared a uint16_t k-mer against 1 << 16, which
    the type already guarantees; a static_assert says what was meant.
  - test_kix_io, test_kpx_io and test_index_filter_validate named
    ikafssn_index / ikafssn_io once through add_ikafssn_test() and again
    through ikafssn_test_index_writer, and the search-helper tests
    re-listed libraries ikafssn_search / ikafssn_builder already carry
    PUBLIC.  Apple's linker reports those as duplicate libraries.

Left alone: the NCBI C++ Toolkit, Drogon and parasail warnings, which
come from upstream sources we only build.

+++ test/test_parallel_sort.cpp
@@ -71,7 +71,9 @@ void test_random_sizes() {
         std::vector<TempEntry> v;
         v.reserve(n);
         for (std::size_t i = 0; i < n; ++i) {
-            v.push_back({rng() & 0xFFFFFFu, rng() & 0xFFFFu, rng() & 0xFFFFu});
+            v.push_back({static_cast<uint32_t>(rng() & 0xFFFFFFu),
+                         static_cast<uint32_t>(rng() & 0xFFFFu),
+                         static_cast<uint32_t>(rng() & 0xFFFFu)});
         }
         char label[64];
         std::snprintf(label, sizeof(label), "rand_n=%zu", n);
@@ -117,7 +119,9 @@ void test_large_random() {
     std::vector<TempEntry> v;
     v.reserve(1 << 17);  // 128K entries
     for (std::size_t i = 0; i < (1u << 17); ++i) {
-        v.push_back({rng() & 0xFFFFFFu, rng() & 0xFFFFu, rng() & 0xFFFFu});
+        v.push_back({static_cast<uint32_t>(rng() & 0xFFFFFFu),
+                     static_cast<uint32_t>(rng() & 0xFFFFu),
+                     static_cast<uint32_t>(rng() & 0xFFFFu)});
     }
     run(std::move(v), "large_rand_128K");
 }
